+++ src/main/java/racingcar/engine/model/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.engine.model;
 
 import java.util.Objects;
