+++ src/test/java/racingcar/RefereeTest.java
@@ -4,6 +4,8 @@ import java.util.Arrays;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingcar.engine.model.Car;
+import racingcar.engine.model.Referee;
 
 public class RefereeTest {
